Changed graph_parse() to require the input graph already to have been transitively reduced.
Formerly, graph_parse() performed a transitive reduction on its input
graph.  This new graph was stored in a temporary object that was
unsuitable for storing in a clanid object; so, the clanid objects
stored a pointer to the original input graph instead.  With the
possibility of recursive calls to graph_parse(), this becomes
unworkable, since the input G will itself often be a temporary created
by an instance of graph_parse() further up the stack.  We need to
distinguish explicitly between the graph being analyzed (possibly a
temporary) and the graph providing the master topology.  Once we've
done that, it makes sense to use subgraphs of the reduced graph for
the recursive calls, which will obviate the need for the transitive
reduction further down the stack.  However, once you do that, you
might as well require the transitive reduction in the original input
to the first call.  So, here we are.