Refactored grain_title so that it works for nodeid_t other than string.
grain_title() now selects the topologically senior node and passes it,
along with the nodeset to unique_nodetitle().  The basic template for
unique_nodetitle() assumes the node type has a member function called
gen_unique(), which it calls to generate a unique node title (nodeset
is ignored).

The std::string specialization of grain_title appends the number of
elements in the nodeset to the name of the best node and returns that
as a string.  Thus, using a string as a nodeid_t will result in
slightly more descriptive grain names (suitable for generating graphs
that show the grain graph).

Other types of nodeid_t should either use the generic version
(implementing the gen_unique() method) or add another template
specialization.