Further optimizations on graph-parse.hh and bitvector.hh
The code now parses the gcam3-noimplied-preproc.dot graph in just under 11.6
seconds on an Evergreen compute node (slightly slower on a login node because
of the additional activity there).

There are probably still some gains to be had, even before considering
parallelizing the graph parsing.  Notably, we recompute the ancestor and
descendant tables on recursive calls (and reallocate the table itself), even
though it only changes a little on recursive calls.

We also spend a lot of time looking nodes up by name, iterating over STL
sets, and so on, particularly in our DFS routines.  If we ever need to
squeeze that out, we can work out a way to refer to nodes strictly by
index and to store their successors/backlinks in a bitvector structure.
This, however, will require some significant reworking of the digraph class,
so we should pursue it only once we have the model run time down below
20s or so.