Fix failure on ComplexGraph test.
Failure was caused by a bogus linear clan being inserted on certain
primitive graph reductions.  We now locate and remove that graph when
appropriate.

The code now must be compiled with the c++-11 standard because we use
an anonymous function in the find() algorithm for the clan we're
looking to remove.  Long-term, we probably want to replace most of our
algorithm predicate classes with anonymous functions as a general code
cleanup measure.