Fix failing GraphParseTest.primitiveReduction test.
This test was failing under the new version of the parser because when
testing the "no side-edges" condition, it assumed that we had
explicitly completed the edges between all of the clan's source nodes
and the union of their children.  Now that we don't explicitly create
those edges the appropriate condition is that all of the clan's
inbound edges come through the source nodes, which is what we test.
Similar adjustments were made to the tests for outbound edges.

I also took this opportunity to clean up some dead code and make some
small documentation fixes.