Adjust naming convention for bitvector class methods.
Previously the bitvector class used size() to return the total number
of elements in the vector, and count() to return the number of
elements that were set.  This is consistent with the way other vector
classes use the word "size".

In practice, however, we invariably use the bitvector to represent a
set, and for a set, size() is supposed to represent the number of
items in the set, which is what count() returns.  This confusion was
particularly problematic when refactoring an STL set into a bitvector
set.  The size() method still existed, but it gave the "wrong" answer.
As a compromise between the two possible uses, we have removed the
size() method entirely.  Now we use length() for the length of the
vector and retain count for the number of set bits.  If you refactor a
conventional container class and forget to change the ambiguous size()
method, you will get a compiler error instead of wrong answers, which
is an improvement.

I also took this opportunity to rename a misnomer in the macros that
set up the popcount table.  It doesn't really affect anything, since
the macros are undefed right after they are used, but it had been
bugging me for a while.