Add GCAM-4 data flow graph to data files.
The GCAM-4 graph is quite a bit more complex than that of GCAM-3.
Apart from the ~50% higher node count, the international trade nodes
tie the regions together in a way that forces us to recurse deeper
than we had to in the GCAM-3 data.  As such, it makes a good stress
test for the parser.