Remove version of digraph::collapse_subgraph() that takes a bitvector argument
Removed because it was misleading.  Collapsing a subgraph changes the
topology of the graph we perform the operation on.  In fact, since it
groups a set of nodes under a single node, it changes the number of
nodes in the graph.  The elements of a set represented by a bitvector
are indexed by their topological indices in the graph they were drawn
for; therefore, this change in topology invalidates all bitvectors
associated with the graph.  In practice, we have to keep a pristine
copy of the original graph so we can keep track of the original
topology.  Having a version of the collapse_subgraph() method that
took a bitvector as an argument created confusing by suggesting that
you could perform this action repeatedly without referring back to the
original topology.

This confusion is related to the problem described in issue #1.  If we
want to do this right we need to separate the set concept from the
underlying representation.